wip: move data models to separate file
