--- .coverage
@@ -1,1 +1,0 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"/home/tyler/pyshipstation-dev/.tox/py375/lib/python3.6/site-packages/shipstation/__init__.py":[1],"/home/tyler/pyshipstation-dev/.tox/py375/lib/python3.6/site-packages/shipstation/api.py":[1,2,3,4,5,8,9,15,28,31,32,33,34,35,55,57,59,61,66,74,80,83,86,94,102,103,108,109,116,122,131,141,153,159,168,181,195,199,206,215,218,250,258,262,268,274,280,286,292,298,301,304,315,321,324,327,334,340,353,356,375,378,397,402,405,410,418,426,384,386,389,391,392,393,395,221,222,223,224,225,228,229,230,231,232,233,234,235,236,237,238,239,240,241,242,243,244,245,246,247,248,62,63,67,72,64,87,92,37,38,39,40,41,43,45,47,49,51,48,44,46,50,75,76,78,95,16,18,19,11,12,13,20,23,21,25,97,84,99,68,69,88,89,70,90,328,332,341,343,325,344,293,296,345,281,284,346,269,272,347,305,306,322,307,310,313,348,335,336,350,444,445,447,448,450,451,452,453]}}
