--- Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pytest = "*"
-pytest-cov = "*"
-coverage-badge = "*"
-nose = "*"
-
-[packages]
-requests = "*"
-
-[requires]
-python_version = "3.7"
