+++ tests/test_customs_item.py
@@ -1,7 +1,8 @@
-import unittest
-import pytest
-from shipstation.api import *
 from decimal import Decimal
+import pytest
+from shipstation.models import *
+from shipstation.api import ShipStation
+import unittest
 
 
 class ShipStationTests(unittest.TestCase):
